Move queue.py to internal.
